Renamed the file name to match the name of the function
