+++ src/app/lib/utilities/preprocessor/rules/filterGroupsWithoutStraplines.js
