+++ src/app/routes/getInitialData/frontpage/index.test.js
@@ -4,7 +4,7 @@ import fetchData from '../utils/fetchData';
 import filterUnknownContentTypes from '../../../lib/utilities/preprocessor/rules/filterContentType';
 import filterEmptyGroupItems from '../../../lib/utilities/preprocessor/rules/filterEmptyGroupItems';
 import applySquashTopstories from '../../../lib/utilities/preprocessor/rules/topstories';
-import filterGroupsWithoutStraplines from '../../../lib/utilities/preprocessor/rules/filterStraplines';
+import filterGroupsWithoutStraplines from '../../../lib/utilities/preprocessor/rules/filterGroupsWithoutStraplines';
 
 const preprocessorRules = [
   filterUnknownContentTypes,
